feat: updated logic to have login sso
