+++ backend/src/services/user/user-service.ts
@@ -151,7 +151,7 @@ export const userServiceFactory = ({
     if (duplicatedAccounts.length && myAccount) {
       await userDAL.transaction(async (tx) => {
         await userDAL.delete({ $in: { id: duplicatedAccounts?.map((el) => el.id) } }, tx);
-        await userDAL.updateById(userId, { username: myAccount.username.toLowerCase() }, tx);
+        await userDAL.updateById(userId, { username: (myAccount.email || myAccount.username).toLowerCase() }, tx);
       });
     }
   };
