+++ frontend/src/hooks/api/users/queries.tsx
@@ -37,15 +37,18 @@ export const useGetUser = () =>
     queryFn: fetchUserDetails
   });
 
+export const fetchUserDuplicateAccounts = async () => {
+  const { data } = await apiRequest.get<{
+    users: Array<User & { isMyAccount: boolean; devices: { ip: string; userAgent: string }[] }>;
+  }>("/api/v1/user/duplicate-accounts");
+  return data.users;
+};
+
 export const useGetMyDuplicateAccount = () =>
   useQuery({
     queryKey: userKeys.getMyDuplicateAccount,
-    queryFn: async () => {
-      const { data } = await apiRequest.get<{
-        users: Array<User & { isMyAccount: boolean; devices: { ip: string; userAgent: string }[] }>;
-      }>("/api/v1/user/duplicate-accounts");
-      return data.users;
-    },
+    staleTime: 60 * 1000, // 1 min in ms
+    queryFn: fetchUserDuplicateAccounts,
     select: (users) => ({
       duplicateAccounts: users.filter((el) => !el.isMyAccount),
       myAccount: users?.find((el) => el.isMyAccount)
